fixed ratingHelper and did % of people rec helper
