+++ client/src/components/reviews/helpers/percentageHelper.js
@@ -1,0 +1,7 @@
+function percentageHelper(recs) {
+  const total = Number(recs.true) + Number(recs.false);
+
+  return Math.round((Number(recs.true) / total) * 100);
+}
+
+export default percentageHelper;
