+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "test": "jest",
     "start": "nodemon server",
-    "react-dev": "webpack --mode development --config webpack.config.js -w & live-server"
+    "react-dev": "webpack --config webpack.config.js -w"
   },
   "repository": {
     "type": "git",
